+++ engine/apps/user_management/models/organization.py
@@ -133,6 +133,8 @@ class Organization(MaintainableObject):
 
     gcom_token = mirage_fields.EncryptedCharField(max_length=300, null=True, default=None)
     gcom_token_org_last_time_synced = models.DateTimeField(null=True, default=None)
+    gcom_org_contract_type = models.CharField(max_length=300, null=True, default=None)
+    gcom_org_has_irm_sku = models.BooleanField(default=False, null=True)
 
     last_time_synced = models.DateTimeField(null=True, default=None)